Setting computed column definition should wait for all property changes (#5613)
The ComputeColumnWindow is initialized from a Json string. This string gets all information to build the drag-and-drop QML window. The signal that this Json string has changed should be emitted when all changes has been made: this is important especially when the chosen column is changed and all properties are suddenly changed. So we must wait for all properties to be changed, before emitting the json change: this is done by connecting the signal with QueuedConnection.