remove WAL from sqlite to avoid bugs
also log how long it took to load or synch a file

Why remove WAL?
It was faster yes, but often the engine would not have the up-to-date info and then crash.
Adding walcheckpoints all over the place didnt seem to help

the slowdown is a bit noticeable with >3000k columns, but given that we didnt support them before shouldnt be a big problem.

For speed comparisons:

        WAL     NO-WAL
3KC      4s        10s
10KC    28s        54s
200KR   30s        30s
1MR    146s       147s
1KR5KC 347s       356s

Where the "R" sets have 2 columns and the "C" sets have a single row.
As can be seen, there *is* a delay for 1K cols and above, but it seems constant.
And any fixes for the synching-db problem with the engine most likely includes adding new delays...
So lets just turn it off