small fix for replacingDoubles on label value change
if there are no 'Label's yet and the user changes the value in the levelseditor to a string some labels will need to be made.
There are some clearly marked functions for this, however, it would always return the last double replaced.
This is usually what is wanted because the other values need not be replaced
but when everything is replaced for convenience the right one should be taken (the one the user edited...)
Not the last one.