Fix filter crash
A bug happened where if you entered in new data:
(one by one)
column1, "c", "1"
column2, "2"

the engine would crash because of some bad housekeeping of data...

this is because labelsAdd was used which checks the valDisMap and then doesnt really add anything.
The db was iterating over rows though...
But ive now replaced it with column->labelsSet.
It takes the honor of either adding a label or overwriting one on a particular place.
And adds them to the also freshly cleaned valDisMap and intsIdMap.
Also, try not to resize the labels to be bigger then it needs to be to avoid all those 0pointer exceptions 😿