Order of variables should be kept in formula
Fixes https://github.com/jasp-stats/INTERNAL-jasp/issues/2689

Also in the Mixed Models, if several random effects are added, some effects are sometimes suddenly added or removed when applying the R Syntax. This was due to the fact that the controls dynamically created by the ComponentsList were not removed before setting the new options.
At last, avoid generating R Syntax during the initialization of the form: each time a control gets a new value, the R Syntax is generated: this is not needed when the form is being initialized.