Show the right implicit empty values in tooltip
If you set a nominal or ordinal variable in a scale VariablesList, a tooltip tells you how the values are transformed into scale.
It gives also a list of 'implicit' empty values: that is the not empty values that could not be converted into a double value. This list was not correct, it gave all the convertable values.