Add types in FactorForms widget
The FactorForms does not specify the types of its variables in the options.  So this information is lost. But this information is needed in the CFA analysis, since it is possible in this analysis to change the variable type.
Also if you change the type, the analysis is not rerun.