QoL
Somehow sandbox got enabled on my mac, but this can never be good
So always false when not windows

Also the message when an analysis has no results yet is a bit misleading as the module actually doesnt take so much time to load.
Only the data really
or the engine starting if black magic happens?