Still a problem when Radiobuttons are created and deleted dynamically
Fixes https://github.com/jasp-stats/jasp-test-release/issues/2911

The RadioButtons are in a TabView. If one tab is removed, the radio buttons are removed: if a radio button is removed before the radio button group is removed, this could provoke an update of the options (the Tab which is deleted, gets a signal to update its option). To tackle this, in the rowControl which contains the controls of the Tab, we call all its controls to be unitialized before deleting them (the blockSignals is not enough since the Radio button can call the Radio Button group without having to use signals).