Solve crash when opening help for Classical Meta-Analysis
Fixes https://github.com/jasp-stats/jasp-issues/issues/3665

A TextArea is a JASPListControl because it may has a model (TextTypeSource, TextTypeJAGSmodel, TextTypeLavaan, TextTypeMetaSem, TextTypeCSem use a model), but other types like the TextTypeDefault one, does not have a model.
So JASPListControl should always check that this model exists, before using it.