Interaction in Factors Form should be kept
In Linear Regression, if you add an interaction between 2 variables in one model, If you duplicate this analysis (or save in a JASP file, and reload this JASP file), then this interaction disappears.

A FactorsForm is a control that has itself an Available VariablesList and 1 or more Assigned VariablesLists. But only the FactorsForm has an option value (is bound), the others Assigned VariablesList do not have their proper option value: their values contribute to the value of the FactorsForm option. In the code, each control that is bound to an option value has a BoundControl object (some controls like Available VariablesList do not have a BoundControl as their value is not set in an option).

The error was that the Assigned VariablesList inside a FactorsForm had still a BoundControl object, and this lead to some confusion in the code.

What happened in Linear Regression is a bit tricky: the "model 1" has the assignAvailableVariablesToList property set to true, which means that when new variables are set in the Available VariablesList, these variables should be automatically set in the "model 1" VariablesList. But when this analysis is initiailized (either by duplicating the analysis, or by loading a JASP file containing a Linear Regression analysis), all controls are initialized when they are created: the "model 1' control is initialized with a null value (as it does not have its own value). In this particular case (assignAvailableVariablesToList is true and no value is given), as there might be default values in the Available VariablesList, which must be set in the Assigned VariablesList (this is for example the case in the ModelTerms of the Repeated Measures ANOVA), this default values are also set in the Assigned VariablesList (line 77 of VariablesListBase.cpp)
This is solved by just testing whether the control has a bound value
Also the BoundControl objects should not be created when the control has no bound value.