Build JASP with Qt Static libs
Only Common, CommonData, QMLComponnents, R-INterface and SyntaxBridge needs to be compiled with static Qt libraries
For this detect this by loading the Qt Core library and set USE_QT_STATIC_QT_LIBS  to ON if the QtCore lib is static.
The SyntaxBridge is built as a shared library so that it can be used directly by the jaspSyntax R package: for this it uses the qt_import_qml_plugins CMAKE function so that Qt loads automatically the needed (static) plugins in the SyntaxBridge library (with the use of a Dummy.sql that uses all the needed import statement).
Also SyntaxBridge needs to load the Qt Minimal plugin platform so that the jaspSyntax can start a Qt application to load the QML form.